requirements.txt
PyGetWindow
pyautogui
Pillow
opencv-python
numpy